Removed some other dependencies
